refactor: consolidate binary crate sprawl into commands/
The binary crate had 3 top-level modules that belong with commands:

- tty.rs → commands/utils/tty.rs (terminal I/O, only used by commands)
- output/ → commands/utils/response.rs (CLI response formatting, only used by commands)
- docs/ → help_topics/ (resolves the 3-way 'docs' name collision:
  src/docs/, commands/docs.rs, and core/code_audit/docs.rs)

After: src/ contains only main.rs, lib.rs, commands/, core/, and
help_topics/. All CLI infrastructure lives under commands/.

771 tests pass, cargo fmt clean, zero new warnings.
