+++ src/main.rs
@@ -16,11 +16,9 @@ enum RawOutputMode {
 }
 
 mod commands;
-mod docs;
-mod output;
-mod tty;
+mod help_topics;
 
-use commands::utils::{args, entity_suggest};
+use commands::utils::{args, entity_suggest, response as output, tty};
 use commands::{
     api, audit, auth, build, changelog, changes, cli, component, config, db, deploy, extension,
     file, fleet, git, init, lint, logs, project, refactor, release, server, ssh, status, test,
