--- src/output/mod.rs
@@ -1,3 +1,0 @@
-mod response;
-
-pub use response::*;
